Ajout et suppression des albums

+++ AlbumPhoto/src/controleur/Album.java
@@ -2,7 +2,10 @@ package controleur;
 
 import java.io.IOException;
 import java.util.Date;
+import java.util.List;
 
+import javax.persistence.EntityManager;
+import javax.persistence.Query;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -13,15 +16,16 @@ import dao.DAOFactory;
 import extra.Data;
 import modele.Personne;
 
-@WebServlet(urlPatterns = { "/Album", "/Album/*" })
+@WebServlet(urlPatterns = {"/Album", "/Album/*", "/Album_delete/*"})
 public class Album extends HttpServlet {
 	private static final long serialVersionUID = 1L;
 	public static final String MESSAGE_DELETE_200 = "L'album a été supprimé avec succès";
 	public static final String MESSAGE_DELETE_404 = "Une erreur est survenue. L'album n'existe pas.";
+	public static final String MESSAGE_DELETE_403 = "Une erreur est survenue. Vous n'êtes pas autorisé à supprimer cet album.";
 	public static final String MESSAGE_POST_UTILISATEUR_NULL_404 = "Cet utilisateur n'existe pas.";
 	public static final String MESSAGE_POST_200 = "L'album a été créé";
 	public static final String MESSAGE_POST_UTILISATEUR_DIFF_404 = "Vous n'avez pas les droits pour modifier cet utilisateur.";
-	public static final String MESSAGE_POST_ATTRIBUTE_MISSING = "Le titre n'a pas été rempli.";
+	public static final String MESSAGE_POST_ATTRIBUTE_MISSING = "le champs titre n'a pas été rempli";
 	public static final String MESSAGE_DELETE_ATTRIBUTE_MISSING = "L'id n'est pas renseigné.";
 	
 	public Album() {
@@ -29,33 +33,19 @@ public class Album extends HttpServlet {
 	}
 
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		// Personne personne = new Personne("Myriam", "Abdel-Fattah");
-		// int idPersonne =
-		// DAOFactory.getInstance().getPersonneDao().create(personne).getId();
-		// DAOFactory.getInstance().getAlbumDao().create(new modele.Album("Mon
-		// ALBUM", new Date(), personne));
-
-		// request.setAttribute("album", new modele.Album("sdfdf", new Date(),
-		// personne));
-
-		// doPost(request, response);
-		String path = request.getPathInfo();
-		if (path == null || path.compareTo("/") == 0) {
-			request.setAttribute("typeRecherche", "album");
-			this.getServletContext().getRequestDispatcher("/vue/rechercher.jsp").forward(request, response);
-		} else {
-			try {
-
-				int idRessource = Integer.parseInt(path.substring(1));
-
-				if (idRessource == 123) {
-					request.setAttribute("album", Data.getAlbumTest());
-					this.getServletContext().getRequestDispatcher("/vue/album.jsp").forward(request, response);
-				} else
-					this.getServletContext().getRequestDispatcher("/vue/ressourceIntrouvable.jsp").forward(request, response);
-			} catch (NumberFormatException ex) {
-				this.getServletContext().getRequestDispatcher("/vue/ressourceIntrouvable.jsp").forward(request, response);
+		String path = request.getPathInfo() != null ? request.getPathInfo().substring(1) : "";
+		if(path.equals("")) {
+			EntityManager entityManager = DAOFactory.getInstance().getEntityManager();
+			Query query = entityManager.createQuery("SELECT x FROM Album x");
+			List<modele.Album> results = (List<modele.Album>) query.getResultList();
+			response.getWriter().append("<table border=\"1\"><tr><th>ID</th><th>Titre</th><th>Créateur</th><th>Date création</th><th>Nombre de photos</th></tr>");
+			for(modele.Album album : results) {
+				response.getWriter().append("<tr><td>"+album.getId()+"</td><td>"+album.getTitre()+"</td><td>"+album.getCreateur().getPrenom()+" "+album.getCreateur().getNom()+"</td><td>"+album.getDateCreation()+"</td><td>"+album.getPhotos().size()+"</td></tr>");
 			}
+			response.getWriter().append("</table>");
+		}
+		else {
+			response.getWriter().append("Rien");
 		}
 	}
 
@@ -64,13 +54,23 @@ public class Album extends HttpServlet {
 		String idAlbum = request.getParameter("idAlbum");
 		if(idAlbum != ""){
 			// Récupére l'album passé en paramètre
-			modele.Album album = (modele.Album) DAOFactory.getInstance().getAlbumDao().read(Integer.getInteger(idAlbum));
+			modele.Album album = (modele.Album) DAOFactory.getInstance().getAlbumDao().read(Integer.parseInt(idAlbum));
 			
 			// Si l'album existe
 			if (album != null) {
-				DAOFactory.getInstance().getAlbumDao().delete(album);
-				request.setAttribute("code", "200");
-				request.setAttribute("message", MESSAGE_DELETE_200);
+				Personne personne = ((modele.Utilisateur)request.getSession().getAttribute(Connexion.ATT_USER_SESSION)).getPersonne();
+				
+				if(album.getCreateur().getId() == personne.getId()){
+					personne.getAlbums().remove(album);
+					DAOFactory.getInstance().getAlbumDao().delete(album);
+					
+					request.setAttribute("code", "200");
+					request.setAttribute("message", MESSAGE_DELETE_200);
+				} else {
+					request.setAttribute("code", "400");
+					request.setAttribute("message", MESSAGE_DELETE_403);
+				}
+				
 			} else {
 				request.setAttribute("code", "400");
 				request.setAttribute("message", MESSAGE_DELETE_404);
@@ -79,25 +79,32 @@ public class Album extends HttpServlet {
 			request.setAttribute("code", "400");
 			request.setAttribute("message", MESSAGE_DELETE_ATTRIBUTE_MISSING);
 		}
-		
+		this.getServletContext().getRequestDispatcher("/vue/gestionnaireAlbums.jsp").forward(request, response);
 	}
 
 	@Override
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		// Récupére les paramètre de l'album qui doit être créé
-		String titre = request.getParameter("titre");
+		if(request.getParameter("method") != null) {
+			if(request.getParameter("method").equals("DELETE")) doDelete(request, response);
+		}
+		else {
 		
-		if(titre != ""){
-			Date date = new Date();
-			Personne personne = ((modele.Utilisateur)request.getSession().getAttribute(Connexion.ATT_USER_SESSION)).getPersonne();
-			modele.Album album = new modele.Album(titre, date, personne);
-			DAOFactory.getInstance().getAlbumDao().create(album);
+			// Récupére les paramètre de l'album qui doit être créé
+			String titre = request.getParameter("titre");
 			
-			request.setAttribute("code", "200");
-			request.setAttribute("message", MESSAGE_POST_200);
-		} else {
-			request.setAttribute("code", "400");
-			request.setAttribute("message", MESSAGE_POST_ATTRIBUTE_MISSING);
+			if(titre != ""){
+				Date date = new Date();
+				Personne personne = ((modele.Utilisateur)request.getSession().getAttribute(Connexion.ATT_USER_SESSION)).getPersonne();
+				modele.Album album = new modele.Album(titre, date, personne);
+				DAOFactory.getInstance().getAlbumDao().create(album);
+				
+				request.setAttribute("code", "200");
+				request.setAttribute("message", MESSAGE_POST_200);
+			} else {
+				request.setAttribute("code", "400");
+				request.setAttribute("message", MESSAGE_POST_ATTRIBUTE_MISSING);
+			}
+			this.getServletContext().getRequestDispatcher("/vue/gestionnaireAlbums.jsp").forward(request, response);
 		}
 	}
